refactor: merge decode errors
